requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework
future==0.18.2
pytz==2022.5
serializers==0.2.4
sqlparse==0.4.3
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5